services/wiq/requirements.txt
Flask==1.0.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
gunicorn==19.9.0
flask-debugtoolbar==0.10.1
flask-cors==3.0.7
flask-migrate==2.4.0
flask-bcrypt==0.7.1
hashids==1.3.1
psycopg2-binary==2.7.7
pyjwt==1.7.1
boto3==1.10.27
boto==2.49.0
botocore==1.13.27
requests==2.21.0